+++ src/senaite/storage/setuphandlers.py
@@ -159,7 +159,7 @@ WORKFLOWS_TO_UPDATE = {
                 "guard": {
                     "guard_permissions": "senaite.storage: Transition: Store Sample",  # noqa
                     "guard_roles": "",
-                    "guard_expr": "python:here.guard_store_sample()",
+                    "guard_expr": "python:here.guard_handler('store')",
                 }
             },
             "recover": {
@@ -171,7 +171,7 @@ WORKFLOWS_TO_UPDATE = {
                 "guard": {
                     "guard_permissions": "senaite.storage: Transition: Recover Sample",  # noqa
                     "guard_roles": "",
-                    "guard_expr": "python:here.guard_recover_sample()",
+                    "guard_expr": "python:here.guard_handler('recover')",
                 }
             },
         }