+++ src/senaite/storage/upgrade/v02_06_000.py
@@ -18,10 +18,12 @@
 # Copyright 2019-2024 by it's authors.
 # Some rights reserved, see README and LICENSE.
 
+from bika.lims import api
 from senaite.core.upgrade import upgradestep
 from senaite.core.upgrade.utils import UpgradeUtils
 from senaite.storage import logger
 from senaite.storage import PRODUCT_NAME
+from senaite.storage.setuphandlers import setup_workflows
 
 version = "2.6.0"
 profile = "profile-{0}:default".format(PRODUCT_NAME)
@@ -46,3 +48,25 @@ def upgrade(tool):
 
     logger.info("{0} upgraded to version {1}".format(PRODUCT_NAME, version))
     return True
+
+
+def remove_guard_scripts(tool):
+    """Removes the guard zope scripts in favour of the generic guard_handler
+    """
+    portal = tool.aq_inner.aq_parent
+
+    # re-import our own workflows
+    setup = portal.portal_setup
+    setup.runImportStepFromProfile(profile, "workflow")
+
+    # Setup the custom workflows
+    setup_workflows(portal)
+
+    # Remove the senaite_storage_scripts skin layer
+    scripts_id = "senaite_storage_scripts"
+    skins_tool = api.get_tool("portal_skins")
+    selections = skins_tool.getSkinSelections()
+    if scripts_id in selections:
+        del selections[scripts_id]
+    if scripts_id in skins_tool:
+        del skins_tool[scripts_id]